Move integration test to test_charm.py
